Auto-PR: Extract duplicate deriveStatus into shared utility (closes #318)
