+++ src/screens/events/DynamicEventDetailScreen.tsx
@@ -35,6 +35,7 @@ import { callEdgeFunction } from '../../utils/edgeFunctions';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { Avatar } from '../../components/ui/Avatar';
 import type { Competition, CompetitionConfig } from '../../utils/supabase';
+import { deriveStatus } from '../../utils/competitionStatus';
 import { getCharityById } from '../../constants/charities';
 
 const RUNSTR_LOGO = require('../../../assets/images/icon.png');
@@ -44,19 +45,6 @@ const BATCH_SIZE = 21;
 interface DynamicEventDetailScreenProps {
   navigation: any;
   route: { params: { eventId: string } };
-}
-
-type EventStatus = 'active' | 'upcoming' | 'ended';
-
-function deriveStatus(comp: Competition): EventStatus {
-  const now = Date.now();
-  const start = new Date(comp.start_date).getTime();
-  const endDate = new Date(comp.end_date);
-  endDate.setUTCHours(23, 59, 59, 999);
-  const end = endDate.getTime();
-  if (now < start) return 'upcoming';
-  if (now > end) return 'ended';
-  return 'active';
 }
 
 const ACTIVITY_ICONS: Record<string, keyof typeof Ionicons.glyphMap> = {
